pyproject.toml
[project]
name = "arxiv-paper-summarizer"
version = "0.1.0"
description = "Summarize ArXiv papers using GitHub Models and create GitHub Issues"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feedparser>=6.0.10",
    "azure-ai-inference>=1.0.0b1",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arxiv_fetcher.py", "src/issue_creator.py", "src/main.py", "src/summarizer.py"]

